+++ examples/level.js
@@ -37,11 +37,10 @@ try {
   assert.ok(e.message.indexOf('name') !== -1)
 }
 
-log.trace("no one should see this")
-log.debug("should see this once (on stdout)")
-log.info("should see this twice")
+log.trace('no one should see this')
+log.debug('should see this once (on stdout)')
+log.info('should see this twice')
 log.levels('stdout', INFO)
-log.debug("no one should see this either")
+log.debug('no one should see this either')
 log.level('trace')
 log.trace('should see this twice as 4th and 5th emitted log messages')
-