+++ examples/src.js
@@ -3,24 +3,23 @@
 
 var Logger = require('../lib/bunyan');
 
-var log = new Logger({name: "src-example", src: true});
+var log = new Logger({name: 'src-example', src: true});
 
-log.info("one");
-log.info("two");
+log.info('one');
+log.info('two');
 function doSomeFoo() {
-  log.info({foo:"bar"}, "three");
+  log.info({foo:'bar'}, 'three');
 }
 doSomeFoo();
 
 function Wuzzle(options) {
   this.log = options.log;
-  this.log.info("creating a wuzzle")
+  this.log.info('creating a wuzzle')
 }
 Wuzzle.prototype.woos = function () {
-  this.log.warn("This wuzzle is woosey.")
+  this.log.warn('This wuzzle is woosey.')
 }
 
-var wuzzle = new Wuzzle({log: log.child({component: "wuzzle"})});
+var wuzzle = new Wuzzle({log: log.child({component: 'wuzzle'})});
 wuzzle.woos();
-log.info("done with the wuzzle")
-
+log.info('done with the wuzzle')