+++ examples/unstringifyable.js
@@ -1,7 +1,7 @@
 // See how bunyan behaves with an un-stringify-able object.
-var Logger = require("../lib/bunyan");
+var Logger = require('../lib/bunyan');
 
-var log = new Logger({src: true, name: "foo"});
+var log = new Logger({src: true, name: 'foo'});
 
 // Make a circular object (cannot be JSON-ified).
 var myobj = {
@@ -9,5 +9,4 @@ var myobj = {
 };
 myobj.myobj = myobj;
 
-log.info({obj: myobj}, "hi there");   // <--- here
-
+log.info({obj: myobj}, 'hi there');   // <--- here